Moved mock FOV handler to proper location.
